backend/requirements.txt
Flask
Flask-CORS
numpy
scikit-learn
sentence-transformers
SpeechRecognition
google-generativeai
gTTS
python-dotenv